requirements.txt
click==7.1.2
sec-edgar-downloader==3.0.5
structlog==20.2.0
wheezy.html==3.0.1
wheezy.template==3.0.3
pyyaml==5.3.1
